refactor: Move functions from email to core
